+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.2.0]
+
+* fixed a bug with new flutter version incompatibility
+* added action extensions
+
 ## [0.1.0]
 
 * updated readme.